Restore separate env and include processing to fix env tests
