+++ lib/spack/spack/config.py
@@ -34,8 +34,6 @@ import functools
 import os
 import re
 import sys
-import urllib.parse
-import urllib.request
 from typing import Any, Callable, Dict, Generator, List, Optional, Tuple, Union
 
 import jsonschema
@@ -68,7 +66,6 @@ import spack.schema.view
 import spack.util.spack_yaml as syaml
 import spack.util.url
 from spack.util.cpus import cpus_available
-from spack.util.path import canonicalize_path, substitute_path_variables
 
 #: Dict from section names -> schema for that section
 SECTION_SCHEMAS: Dict[str, Any] = {
@@ -123,9 +120,6 @@ _OVERRIDES_BASE_NAME = "overrides-"
 
 #: Type used for raw YAML configuration
 YamlConfigDict = Dict[str, Any]
-
-#: Track staged paths so can avoid repeating the warning during processing
-skip_restage_warning = list()
 
 
 class ConfigScope:
@@ -400,164 +394,6 @@ class InternalConfigScope(ConfigScope):
                 result[key] = copy.copy(sv)
 
         return result
-
-
-# TODO/TBD/TLD: Do we want/need resolve_relative or can it be inlined?
-def scopes_from_paths(
-    includes: List[str],
-    name_prefix: str,
-    config_stage_dir: str,
-    resolve_relative: Callable[[str], str],
-    required: List[str] = [],
-) -> List[ConfigScope]:
-    """Load included config scopes
-
-    Scopes are added in reverse order so that highest-precedence scopes are last.
-
-    Args:
-        includes: list of paths to be included
-        name_prefix: environment's name prefix
-        config_stage_dir: path to directory to be used to stage remote includes
-        resolve_relative: method to use to resolve relative paths
-        required: list of paths that are required
-
-    Raises:
-        ValueError: included path has an unsupported URL scheme or is required
-            but does not exist
-        spack.config.ConfigError: cannot fetch remote configuration
-    """
-    scopes: List[ConfigScope] = []
-
-    missing = []
-    for i, orig_path in enumerate(reversed(includes)):
-        # allow paths to contain spack config/environment variables, etc.
-        config_path = substitute_path_variables(orig_path)
-        include_url = urllib.parse.urlparse(config_path)
-
-        # If scheme is not valid, config_path is not a url
-        # of a type Spack is generally aware
-        if spack.util.url.validate_scheme(include_url.scheme):
-            # Transform file:// URLs to direct includes.
-            if include_url.scheme == "file":
-                config_path = urllib.request.url2pathname(include_url.path)
-
-            # Any other URL should be fetched.
-            elif include_url.scheme in ("http", "https", "ftp"):
-                # Stage any remote configuration file(s)
-                staged_configs = (
-                    os.listdir(config_stage_dir) if os.path.exists(config_stage_dir) else []
-                )
-                remote_path = urllib.request.url2pathname(include_url.path)
-                basename = os.path.basename(remote_path)
-                if basename in staged_configs:
-                    # Do NOT re-stage configuration files over existing
-                    # ones with the same name since there is a risk of
-                    # losing changes (e.g., from 'spack config update').
-                    if remote_path not in skip_restage_warning:
-                        tty.warn(
-                            f"Will not re-stage configuration from {remote_path} "
-                            "to avoid losing changes to the already staged file of "
-                            "the same name."
-                        )
-                        skip_restage_warning.append(remote_path)
-
-                    # Recognize the configuration stage directory
-                    # is flattened to ensure a single copy of each
-                    # configuration file.
-                    config_path = config_stage_dir
-                    if basename.endswith(".yaml"):
-                        config_path = os.path.join(config_path, basename)
-                else:
-                    staged_path = fetch_remote_configs(
-                        config_path, str(config_stage_dir), skip_existing=True
-                    )
-                    if not staged_path:
-                        raise spack.error.ConfigError(
-                            f"Unable to fetch remote configuration {config_path}"
-                        )
-                    config_path = staged_path
-
-            elif include_url.scheme:
-                raise ValueError(
-                    f"Unsupported URL scheme ({include_url.scheme}) for an "
-                    f"include: {config_path}"
-                )
-
-        # TBD/TLD Is using canonicalize_path equivalent? ..will repeat
-        #   call to substitute_path_variables *after* BUT makes absolute
-        #   path relative to the dir the file is written in.
-        if not os.path.isabs(config_path):
-            config_path = resolve_relative(config_path)
-
-        if os.path.isdir(config_path):
-            # directories are treated as regular ConfigScopes
-            config_name = f"{name_prefix}:{os.path.basename(config_path)}"
-            tty.debug(f"Creating DirectoryConfigScope {config_name} for '{config_path}'")
-            scopes.append(DirectoryConfigScope(config_name, config_path))
-        elif os.path.exists(config_path):
-            # files are assumed to be SingleFileScopes
-            config_name = f"{name_prefix}:{config_path}"
-            tty.debug(f"Creating SingleFileScope {config_name} for '{config_path}'")
-            scopes.append(SingleFileScope(config_name, config_path, spack.schema.merged.schema))
-        elif orig_path in required:
-            raise ValueError(f"Required include path does not exist: {orig_path}")
-        else:
-            missing.append(config_path)
-            continue
-
-    if missing:
-        msg = f"Detected {len(missing)} optional missing include path(s):"
-        msg += "\n   {0}".format("\n   ".join(missing))
-        tty.warn(msg)
-
-    return scopes
-
-
-# TODO/TLD: Todd says can use spack.util.path.canonicalize_path since it
-#   already handles relative path resolution for paths from config files
-def included_config_scopes(
-    name_prefix: str, config_stage_dir: str, resolve_relative: Callable[[str], str]
-) -> List[ConfigScope]:
-    """List of included configuration scopes.
-
-    Args:
-        name_prefix: scope name prefix
-        config_stage_dir: path to directory to be used to stage remote includes
-        resolve_relative: method to use to resolve relative paths
-
-    Returns: included configuration scopes
-    """
-    import spack.environment.environment as env  # circular import
-
-    includes = get("include")
-    if not includes:
-        return []
-
-    include_paths = list()
-    required_paths = list()
-    for entry in includes:
-        always_activate = False
-        optional = False
-        if isinstance(entry, str):
-            path = entry
-            always_activate = True
-        else:
-            path = entry["path"]
-            if "when" in entry:
-                when_str = entry["when"]
-            else:
-                always_activate = True
-            optional |= entry.get("optional", False)
-
-        if not optional:
-            required_paths.append(path)
-
-        if always_activate or env._eval_conditional(when_str):
-            include_paths.append(path)
-
-    return scopes_from_paths(
-        include_paths, name_prefix, config_stage_dir, resolve_relative, required_paths
-    )
 
 
 def _config_mutator(method):
@@ -954,8 +790,6 @@ def update_config_with_includes():
     configurations to include. This does not handle recursive includes
     (i.e. if an included config defines an "include:" section).
     """
-    import spack.environment.environment as env  # circular import
-
     includes = CONFIG.get("include")
     if not includes:
         return
@@ -975,20 +809,20 @@ def update_config_with_includes():
                 always_activate = True
             optional |= entry.get("optional", False)
 
-        include_path = canonicalize_path(include_path)
+        include_path = spack.util.path.canonicalize_path(include_path)
         if not os.path.exists(include_path) and not optional:
             raise ValueError(
                 f"Specified include path does not exist and is not optional: {include_path}"
             )
 
-        activate = always_activate or env._eval_conditional(when_str)
+        activate = always_activate or spack.environment.environment._eval_conditional(when_str)
         if activate and os.path.exists(include_path):
             to_add.append(include_path)
 
     def resolve_relative(cfg_path):
         raise ValueError(f"config:include got relative path {cfg_path}")
 
-    scopes = scopes_from_paths(
+    scopes = spack.environment.environment.scopes_from_paths(
         to_add, "include", config_stage_dir=None, resolve_relative=resolve_relative
     )
